Fix cluster tag not being sent to new worker applications
Fixes lp:1824361

When using CDK with the AWS integrator, if a new worker application is
added to a running cluster, the masters will never send the cluster tag
to the new workers, which prevents the cloud integration with the new
workers from completing. They workers will get stuck in a "Waiting for
cloud integration" status.

This change removes the state that says, "I've already sent the cluster
tag, never send it again," thereby allowing the cluster tag to be sent
to new worker applications that join the cluster.