Report in status if Vault encryption is expected but not ready (#56)
Due to https://bugs.launchpad.net/charm-kubernetes-master/+bug/1843809
we hit an issue where Vault was related for secrets encryption but never
provided the config needed to actually do the encryption. However, this
was never indicated in any way which would lead the admin to expect that
the secrets would be encrypted at rest when they would not in fact be.