Remove broken legacy-handling of client-ca-file apiserver arg (#62)
Once upon a time this code worked properly, because each api-server arg
was passed to 'snap set' separately, and args that had a value of 'null'
would be unset, and therefore not passed as an arg when the api-server
was started.

Since then, we've changed the snap such that all args to the api-server
are passed to the snap via a single 'snap set args="..."'. Instead of
causing an arg to be unset, a value of 'null' will now be passed through
to the api-server, which in this case, causes the api-server to fail to
start.

Removing client-ca-file from api_opts when it's not needed fixes the
problem. If it's not there, it won't be passed as an arg to the
api-server.

Fixes: https://bugs.launchpad.net/charm-kubernetes-master/+bug/1845962