Handle upgrades for sending new cluster-name to cdk-addons
This fixes the issue where upgrading the charm code will cause the
OpenStack cloud provider to use a new `cluster-name` value and thus
recreate all load balancers with new names, requiring manual
intervention. This makes it so the operator needs to invoke the upgrade
action to enable the new cluster name and can explicitly prevent the
switch by overriding the `fix-cluster-name` action param.