Changes after review and reflection
Backed away from removing the policy configmap when not using authorization due to upstream docs suggesting to keep it to prevent bouncing the pods when it is turned on or off.
Added code to remove policy when keystone relation is removed.
Created error flag to surface error messages to the user in the case of the policy file failing to apply or not being applied yet.