Send fsName to cdk-addons for CephFS
The name of the CephFS file system created by the ceph-fs charm is not
hard-coded. Instead, it is set to the application name that the ceph-fs
charm was deployed with, which we cannot determine via Juju because we
have no direct relation to ceph-fs. So we have to query ceph to find the
correct value to use for fsName.