Write and pass in kubeconfig for cdk-addons
It seems that cdk-addons has always expected a kubeconfig snap config to
be passed in with a path to the kubectl config file to use, but
k8s-master has never actually created and set that. It happened to work
by accident up until now, since kubectl would try localhost:8080 which
happened to work. It seems the latest kubectl getting built into
cdk-addons no longer falls back to localhost:8080 (see k8s PR#86173), so
this uncovered the fact that we were not giving an explicit config.