support for service-cidr expansion (#93)
kubernetes-master charm does not allow service-cidr to be changed
once k8s is deployed. However k8s by itself supports service-cidr
expansion. This PR adds support for service-cidr expansion in
kubernetes-master charm.

The charm checks whether the new service-cidr is an expansion or not.
If it is not an expansion, a warning message is displayed on the
status line.

For expansion case,
* regenerate API certs
* update kube-apiserver parameters
* restart kube-apiserver apiservice
* delete kubernetes service object
* rollout cdk-addon services

Closes-Bug: #1868685

Co-authored-by: Hemanth Nakkina <hemanth.nakkina@canonical.com>