move peer logic where its needed: before sending worker cohorts
It's possible that master peers are coming (goal-state says so), but the
peer relation hasn't fired yet. Masters do need to eventually agree, but
only prior to sending cohort data over the kube-control worker relation.

Requiring the agreement any earlier delays the deployment of master
followers because they're stuck in a reactive loop waiting for the
peer relation to become ready so they can set 'cohorts.joined'.

It's safe (because of layer-coordinator) for all master units to join
a cohort at any time. As noted, the only place we *need* master agreement
is just before sending cohort data to workers.