Don't require full Keystone creds from relation
The charm actually only cares about the auth API URL but because there's
not a relation for just that from Keystone, it instead uses the relation
for requesting a new user (keystone-credentials). The interface layer
does provide a flag for the case where just the API endpoint info is
available, so it's better to use that instead of blocking on the full
credentials being available.

This fixes part of [lp:1834164][] where the integrator charm might not
have credentials in the format required by the relation but can still at
least provide the API endpoint info.

[lp:1834164]: https://bugs.launchpad.net/charm-kubernetes-master/+bug/1834164