Prevent race conditions for config creation
Some of the kubeconfig files created in the
build_kubeconfig() function need the related
secret to be already present in the cluster.
The creation of such secrets is managed in the
create_tokens_and_sign_auth_requests() function,
called in the setup_leader_authentication() function.

This last function needs `authentication.setup` to
be unset, but does not explicitly need the apiserver
to be configured. Nevertheless, the setup_tokens()
function does not create a secret if the apiserver
is not configured. This can lead to race conditions.

Explicitly calling again the build_kubeconfig() and the
create_tokens_and_sign_auth_requests() function in the
setup_auth_webhook_tokens(), i.e. after the apiserver
configuration, ensures that all secrets are created
and all configuration files are saved, preventing race
conditions and the consequent missing files errors.