Fix Focal deploys broken by Werkzeug 2.0
Yesterday (May 11), Werkzeug 2.0.0 was released which includes a
conditional dependency of dataclasses for Python <3.7 but that fails to
install on 3.7+. Because the charm snap uses Python 3.6 during the build
but Focal includes Python 3.8, deployments there are now broken.

This pins the Werkzeug library and adds tests to validate that the built
wheelhouse will successfully install on a given Python version.

Fixes: [lp:1928254][]

[lp:1928254]: https://bugs.launchpad.net/charm-kubernetes-master/+bug/1928254