CK-124 - Fix port for loadbalancer-external
The old external LB support using kubeapi-load-balancer and the old
public-address interface used the default config value for the port
from the ka-l-b charm of 443. The switch to the new loadbalancer
interface ended up unintentionally switching that in the default
configuration to use the standard API port of 6443 which broke the
test_sans test case.