Ensure charm only accesses local API server
The new LB support in #153 unintentionally changed the endpoint that the
charm client uses to talk to the API server making it use the internal
LB address instead of always talking to the API server locally. This
introduced an issue during bootstrap where the initial local-only token
would try to be used on other API servers and fail.

Fixes [lp:1941763](https://bugs.launchpad.net/charm-kubernetes-master/+bug/1941763)