HACluster - Removing master services from Pacemaker and leaving to Systemd (#190)
Currently, charm-hacluster is not fully prepared to receive kubernetes master services
for some reasons:

1. charm-hacluster doesn’t wait to receive all constraints of order and colocation necessary to
run master services without conflict. Kube-apiserver should be running in a node before
Pacemaker tries to start kube-controller-manager in the same node. [1]

2. Master services are cloned (running in all nodes) with Pacemaker and not grouped with VIP
resources, meaning that if a resource starts to fail in a node that hosts the VIP, it won’t try to
change node leaving the VIP in a unit where it’s not healthy. [2]

3. Resources added with the hacluster-interface have the `migration-threshold` set to
`INFINITY`. This means that the cluster will try to recover the service in the node forever. In a
scenario where just restarting the service is not able to recover the service, Pacemaker won’t
change the VIP to another node even if it’s grouped with the master services. [3]

Knowing those issues, the short fix proposed in this PR is to remove the master services from
Pacemaker when hacluster is present in the bundle and leave to systemd that will try to restart if
the service fails. The `atexit` checks the health of master services and if some service is not
running, it will set the node in `standby` mode and this will force migration of VIP to another
node.

To make those changes take effect, first it’s necessary to approve the layer-hacluster PR [4].

One known problem of this approach is that sometimes kubernetes-master hooks get "stuck" in
a busy loop and never exit which would affect the logic needed in `atexit`. Besides this, it could
take up to 5 minutes to identify the problem and change the VIP to another node in the cluster.
The long-term solution needs to solve problems 1, 2 and 3.

[1] https://bugs.launchpad.net/charm-hacluster/+bug/1952492
[2] https://bugs.launchpad.net/charm-interface-hacluster/+bug/1952753
[3] https://github.com/openstack/charm-interface-hacluster/blob/master/interface_hacluster/common.py#L993-L1009
[4] https://github.com/juju-solutions/layer-hacluster/pull/3