Fix dashboards to relate the cluster and juju model dropdowns (#397)
When the dashboards are received by the COS grafana charm, it adds
additional dropdowns with juju topology information (model, uuid,
application, etc.) These added dropdowns are not related to the
dropdowns sent by this charm. This patch adds a reference to the juju
model on the cluster dropdowns so that they only show the clusters
belonging to the selected juju models.

Signed-off-by: Jorge Merlino <jorge.merlino@canonical.com>